Cast slider movement as int
Sliders behave strangely otherwise despite the slider itself showing the correct value. Upon using the context menu to a value, it will bring up a textbox with a strange value such as 14326.1 instead of 14300. Casting as int seems to prevent this and just function as intended.

+++ FamiStudio/Source/App/Common/Controls/Slider.cs
@@ -80,14 +80,14 @@ namespace FamiStudio
 
             if (changing)
             {
-                Value = Utils.Lerp(min, max, Utils.Saturate(e.X / (float)(width)));
+                Value = (int)Utils.Lerp(min, max, Utils.Saturate(e.X / (float)(width)));
                 e.MarkHandled();
             }
         }
 
         protected override void OnTouchClick(PointerEventArgs e)
         {
-            Value = Utils.Lerp(min, max, Utils.Saturate(e.X / (float)(width)));
+            Value = (int)Utils.Lerp(min, max, Utils.Saturate(e.X / (float)(width)));
             e.MarkHandled();
         }
